Refactor how we add stimulus controller html attributes so that we can combine them
