+++ app/views/form/_numeric_question.html.erb
@@ -3,5 +3,5 @@
     label: { text: question["header"].html_safe, size: "l"},
     min: question["min"], max: question["max"], step: question["step"],
     width: 20, :readonly => question["readonly"],
-   **numeric_question_html_attributes(question)
+   **stimulus_html_attributes(question)
 %>
