+++ app/views/form/_radio_question.html.erb
@@ -7,7 +7,7 @@
       <% if key.starts_with?("divider") %>
         <%= f.govuk_radio_divider %>
       <% else %>
-        <%= f.govuk_radio_button question_key, val, label: { text: val }, **conditional_html_attributes(question) %>
+        <%= f.govuk_radio_button question_key, val, label: { text: val }, **stimulus_html_attributes(question) %>
       <% end %>
     <% end %>
 <% end %>
