+++ app/views/form/_text_question.html.erb
@@ -2,5 +2,5 @@
     hint: { text: question["hint_text"] },
     label: { text: question["header"].html_safe, size: "l"},
     width: 20,
-    **conditional_html_attributes(question)
+    **stimulus_html_attributes(question)
 %>
