+++ spec/helpers/stimulus_controller_attribute_helper_spec.rb
@@ -1,16 +1,16 @@
 require "rails_helper"
 
-RSpec.describe NumericQuestionsHelper do
+RSpec.describe StimulusControllerAttributeHelper do
   let(:form) { Form.new(2021, 2022) }
   let(:questions) { form.questions_for_page("rent") }
 
   describe "html attributes" do
     it "returns empty hash if fields-to-add or result-field are empty " do
-      expect(numeric_question_html_attributes(questions["total_charge"])).to eq({})
+      expect(stimulus_html_attributes(questions["total_charge"])).to eq({})
     end
 
     it "returns html attributes if fields-to-add or result-field are not empty " do
-      expect(numeric_question_html_attributes(questions["basic_rent"])).to eq({
+      expect(stimulus_html_attributes(questions["basic_rent"])).to eq({
         "data-controller": "numeric-question",
         "data-action": "numeric-question#calculateFields",
         "data-target": "case-log-#{questions['basic_rent']['result-field'].to_s.dasherize}-field",
